Arduino_marriage_proposal.ino
I wrote this code with the assistance of software developer Karolin Ecke to propose to my girlfriend. I surprised my girlfriend one day with
a locked 'treasure hunt' box, with an Arduino inside that powered an LCD screen for clues and a button to change to the next clue.
Each clue led her to a different location in our city where one of her friends was waiting with a key or combination to unlock a part of the box.
She said yes!